requirements.txt
reportlab
pyyaml
psutil
psycopg2
matplotlib
seaborn
pandas
numpy